+++ src/SFA.DAS.ApprenticeCommitments.Api.AcceptanceTests/WorkflowTests/EmailPropertyCustomisation.cs
@@ -1,19 +1,10 @@
 ﻿using AutoFixture;
 using AutoFixture.Kernel;
-using FluentAssertions;
-using NUnit.Framework;
-using SFA.DAS.ApprenticeCommitments.Application.Commands.CreateAccountCommand;
-using SFA.DAS.ApprenticeCommitments.Application.Commands.CreateRegistrationCommand;
-using SFA.DAS.ApprenticeCommitments.Application.Commands.CreateApprenticeshipFromRegistrationCommand;
-using System;
-using System.Net.Http;
 using System.Net.Mail;
 using System.Reflection;
-using System.Threading.Tasks;
 
 namespace SFA.DAS.ApprenticeCommitments.Api.AcceptanceTests.WorkflowTests
 {
-
     internal class EmailPropertyCustomisation : ISpecimenBuilder
     {
         public object Create(object request, ISpecimenContext context)
